requirements.txt
python-telegram-bot
requests
Pillow
alphabet-detector